+++ packages/swapper/src/swappers/thorchain/types.ts
@@ -1,12 +1,12 @@
 import { ChainId } from '@shapeshiftoss/caip'
 import { ChainAdapterManager } from '@shapeshiftoss/chain-adapters'
-import { BTCSignTx, CosmosSignTx, ETHSignTx } from '@shapeshiftoss/hdwallet-core'
+import type { BTCSignTx, CosmosSignTx, ETHSignTx } from '@shapeshiftoss/hdwallet-core'
 import { KnownChainIds } from '@shapeshiftoss/types'
 import type Web3 from 'web3'
 
-import { Trade, UtxoSupportedChainIds } from '../../api'
+import type { Trade, UtxoSupportedChainIds } from '../../api'
 
-export type PoolResponse = {
+export type MidgardPoolResponse = {
   asset: string
   assetDepth: string
   assetPrice: string
@@ -21,13 +21,28 @@ export type PoolResponse = {
   volume24h: string
 }
 
-type MidardCoins = {
+export type ThornodePoolResponse = {
+  LP_units: string
+  asset: string
+  balance_asset: string
+  balance_rune: string
+  pending_inbound_asset: string
+  pending_inbound_rune: string
+  pool_units: string
+  status: string
+  synth_supply: string
+  synth_units: string
+}
+
+type MidgardCoins = {
   asset: string
 }[]
+
 type MidgardActionOut = {
-  coins: MidardCoins
+  coins: MidgardCoins
   txID: string
 }
+
 type MidgardAction = {
   date: string
   height: string
@@ -50,6 +65,7 @@ export type InboundResponse = {
 }
 
 export type ThorchainSwapperDeps = {
+  daemonUrl: string
   midgardUrl: string
   adapterManager: ChainAdapterManager
   web3: Web3